% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{badouble_declareDependent}
\alias{badouble_declareDependent}
\title{Identify dependent variables}
\usage{
badouble_declareDependent(...)
}
\arguments{
\item{variable(s)}{}
}
\value{
None
}
\description{
\code{badouble_declareDependent} identifies the dependent variables in the computation. 
\code{badouble_declareDependent} can take a single or multiple arguments. 
It is used to identify those output variables that are dependents in computation 
being differentiated. Derivatives will be computed for identified output 
variables with respect to the identified input variables 
(using \code{badouble_declareIndependent}).
}
\examples{
library(radolc)

trace_on(1)
x <- c(adouble(1.0),adouble(2.0))
badouble_declareIndependent(x)
#... calculations.
trace_off(1);

#Always detach the package
detach(package:radolc, unload=TRUE) 

}
\seealso{
{\code{\link{trace_on}}, \code{\link{trace_off}}, 
        \code{\link{badouble_declareIndependent}}}
}
\keyword{{radolc}}
