% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{badouble_declareIndependent}
\alias{badouble_declareIndependent}
\title{Identify independent variables}
\usage{
badouble_declareIndependent(...)
}
\arguments{
\item{variable(s)}{}
}
\value{
None
}
\description{
\code{badouble_declareIndependent} identifies the independent variables in the computation. 
\code{badouble_declareIndependent} can take a single or multiple arguments. 
It is used to identify those input variables that are independents in computation 
being differentiated. Derivatives will be computed for identified output 
variables (using \code{badouble_declareDependent}) with respect to the 
identified input variables.
}
\examples{
library(radolc)

trace_on(1)
x <- c(adouble(1.0),adouble(2.0))
y <- c(adouble(1.0),adouble(2.0))
badouble_declareIndependent(x)
#... calculations.
badouble_declareDependent(y)
trace_off(1);

#Always detach the package
detach(package:radolc, unload=TRUE) 
}
\seealso{
{\code{\link{trace_on}}, \code{\link{trace_off}}, 
        \code{\link{badouble_declareDependent}}}
}
\keyword{{radolc}}
