% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{gradient}
\alias{gradient}
\title{Compute the gradient of the dependent with respect to the independent variable(s)}
\usage{
gradient(s_arg1, s_arg2, s_arg3, s_arg4, .copy = FALSE)
}
\arguments{
\item{tag}{integer for tape identification (used in trace_on())}

\item{n}{integer number of independents n and m = 1}

\item{x}{independent vector of length n}

\item{g}{resulting gradient of length n}
}
\value{
None. the output is the the parameter g
}
\description{
\code{gradient} computes the derivatives for dependent variable with respect to
n independents for the computation that has been traced.
}
\examples{
library(radolc)

fr <- function(x) {   ## Rosenbrock Banana function
    x1 <- x[[1]]
    x2 <- x[[2]]
    y <- 100 * (x2 - x1 * x1)* (x2 - x1 * x1) + (1 - x1)*(1 - x1)
    y }

tag <- 1
trace_on(tag)
x <- c(adouble(1.0),adouble(2.0))
badouble_declareIndependent(x)
y <- fr(x)

badouble_declareDependent(y)
trace_off()
#gradient(tag,n,x,g)
# tag: integer, tape identification
# n  : integer, number of independents n and m = 1
# x[n]: independent vector x
# g[n] :resulting gradient \gradF(x)

x=c(1,2)
g <- c(0.0,0.0)
gradient(1,2,x,g)

#Always detach the package
detach(package:radolc, unload=TRUE)

}
\seealso{
{\code{\link{trace_on}}, \code{\link{trace_off}}, 
        \code{\link{badouble_declareIndependent}},
        \code{\link{badouble_declareDependent}}}
}
\keyword{{radolc}}
