% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{hess_mat}
\alias{hess_mat}
\title{Compute the Hessian matrix product}
\usage{
hess_mat(s_arg1, s_arg2, s_arg3, s_arg4, s_arg5, s_arg6, .copy = FALSE)
}
\arguments{
\item{tag}{integer for tape identification (used in \link{trace_on})}

\item{n}{integer number of independents and m = 1}

\item{x}{independent vector of length n}

\item{v}{vector of length n}

\item{z}{resulting vector of lenth n \nabla^2F(x)

# hess_mat(tag, n, q, x[n], V[n][q], W[n][q]) # result z = \nabla^2F(x)v}
}
\value{
None. the output is the the parameter z
}
\description{
\code{hess_vec} computes the product of Hessain matrix with a matrix of suitable dimension
}
\examples{
hess_mat(tag, n, q, x, V, W)

}
\seealso{
{\code{\link{trace_on}}, \code{\link{trace_off}}, 
        \code{\link{badouble_declareIndependent}},
        \code{\link{badouble_declareDependent}}}
}
\keyword{{radolc}}
