% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{hess_vec}
\alias{hess_vec}
\title{Compute the Hessian vector product}
\usage{
hess_vec(s_arg1, s_arg2, s_arg3, s_arg4, s_arg5, .copy = FALSE)
}
\arguments{
\item{tag}{integer for tape identification (used in \link{trace_on})}

\item{n}{integer number of independents and m = 1}

\item{x}{independent vector of length n}

\item{v}{vector of length n}

\item{z}{resulting vector of length n}
}
\value{
None. the output is the the parameter z
}
\description{
\code{hess_vec} computes the product of Hessian matrix with a vector of suitable size
}
\examples{
rm(list=ls())

library(radolc)

#--- testing ADOLC's hessian vector product
fr <- function(x) {   ## Rosenbrock Banana function
  x1 <- x[1]
  x2 <- x[2]
  y <- 100 * (x2 - x1 * x1)* (x2 - x1 * x1) + (1 - x1)*(1 - x1)
  y }

trace_on(1)
x <- adolc_createList(2,2.0)
badouble_declareIndependent(x)
y <- fr(x)
badouble_declareDependent(y)
trace_off()

xx <- c(1.0,2.0)
vv <- c(3.0,5.5)
yy <- c(0.0,0.0)
hess_vec(1,2,xx,vv,yy) # result z = \nambla^2F(x)v
#hess_vec(tag,n,x,v,z)
# tag: integer, tape identification
# n  : integer, number of independents n and m = 1
# x[n]: independent vector x
# v[n]: vector
# z[n]: resulting Hessian vector product

print(yy)     

#Always detach the package
detach(package:radolc, unload=TRUE) 
}
\seealso{
{\code{\link{trace_on}}, \code{\link{trace_off}}, 
        \code{\link{badouble_declareIndependent}},
        \code{\link{badouble_declareDependent}}}
}
\keyword{{radolc}}
