% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{hessian}
\alias{hessian}
\title{Compute the second derivatives of the dependent variables with respect to the independent variable}
\usage{
hessian(s_arg1, s_arg2, s_arg3, s_arg4, .copy = FALSE)
}
\arguments{
\item{tag}{integer for tape identification (used in \link{trace_on})}

\item{n}{integer number of independents and m = 1}

\item{x}{independent vector of length n}

\item{H}{resulting Hessian matrix (n*n)}
}
\value{
None. the output is the the parameter H
}
\description{
\code{hessian} computes the Hessian matrix values for m dependents with respect to
n independents for the computation that has been traced.
}
\examples{
rm(list=ls())

library(radolc)

#--- testing ADOLC's hessian
fr <- function(x) {   ## Rosenbrock Banana function
  x1 <- x[1]
  x2 <- x[2]
  y <- 100 * (x2 - x1 * x1)* (x2 - x1 * x1) + (1 - x1)*(1 - x1)
  y }

trace_on(1)
x <- adolc_createList(2,2.0)
badouble_declareIndependent(x)
y <- fr(x)
badouble_declareDependent(y)
trace_off()

xx <- c(1.0,2.0)
yy <- matrix(rep(0.0,4), nrow = 2, ncol = 2)
hessian(1,2,xx,yy);
#hessian(tag,n,x,H)
# tag: integer, tape identification
# n  : integer, number of independents n and m = 1
# x[n]: independent vector x
# double H[n][n]: resulting Hessian matrix \nabla^2F(x)
print(yy)     

#Always detach the package
detach(package:radolc, unload=TRUE) 
}
\seealso{
{\code{\link{trace_on}}, \code{\link{trace_off}}, 
        \code{\link{badouble_declareIndependent}},
        \code{\link{badouble_declareDependent}}}
}
\keyword{{radolc}}
