% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{jacobian}
\alias{jacobian}
\title{Compute the derivatives of the dependent variables with respect to the independent variables}
\usage{
jacobian(s_arg1, s_arg2, s_arg3, s_arg4, s_arg5, .copy = FALSE)
}
\arguments{
\item{tag}{integer for tape identification (used in \link{trace_on})}

\item{m}{integer, number of dependent variables}

\item{n}{integer number of independents}

\item{x}{independent vector of length n}

\item{J[m][n]}{resulting Jacobian (m*n) F(x)}
}
\value{
None. the output is the the parameter J
}
\description{
\code{jacobian} computes the Jacobian matrix values for m dependents with respect to
n independents for the computation that has been traced. If m>n then the forward mode of automatic differentiation is used. 
Otherwise, the reverse emode of automatic differentiation is used.
}
\examples{
rm(list=ls())

library(radolc)

#--- testing ADOLC's jacobian

fr <- function(x) {
  x1 <- x[[1]]
  x2 <- x[[2]]
  #Use create list to create adouble objects
  y<-adolc_createList(2,2.0)
  #y <- c(0.0,0.0)
  y[[1]] <- 100 * (x2 - x1 * x1)* (x2 - x1 * x1) + (1 - x1)*(1 - x1)
  y[[2]] <- 100 * (x2 - x1 * x1)+ (x2 - x1 * x1) * (1 - x1)*(1 - x1)
  y}

trace_on(1)
x <- adolc_createList(2,2.0)
badouble_declareIndependent(x)
y <- fr(x)
badouble_declareDependent(y)
trace_off()

xx <- c(1.0,2.0)
yy <- matrix(rep(0.0,4), nrow = 2, ncol = 2)
jacobian(1,2,2,xx,yy);
#jacobian(tag,m,n,x,J)
# tag: integer, tape identification
# m  : integer, number of dependent variables
# n  : integer, number of independent variables n
# x[n]: independent vector x
#J[m][n]; // resulting Jacobian F(x)
print(yy)     

#Always detach the package
detach(package:radolc, unload=TRUE) 
}
\seealso{
{\code{\link{trace_on}}, \code{\link{trace_off}}, 
        \code{\link{badouble_declareIndependent}},
        \code{\link{badouble_declareDependent}}}
}
\keyword{{radolc}}
