% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{gradient}
\alias{gradient}
\title{Compute the gradient of the dependent with respect to the independent variable(s)}
\usage{
gradient(s_arg1, s_arg2, s_arg3, s_arg4, .copy = FALSE)
}
\arguments{
\item{tag}{integer for tape identification (used in trace_on())}

\item{n}{integer number of independents n and m = 1}

\item{x}{independent vector of length n}

\item{g}{resulting gradient (\gradF(x)) of length n}
}
\value{
None. the output is the the parameter g
}
\description{
\code{gradient} computes the derivaties for dependent varible with respect to
n independents for the computation that has been traced.
}
\examples{
x=c(1,2)
g <- c(0.0,0.0)
gradient(1,2,x,g)

}
\seealso{
{\code{\link{trace_on}}, \code{\link{trace_off}}, 
        \code{\link{badouble_declareIndependent}},
        \code{\link{badouble_declareDependent}}}
}
\keyword{{radolc}}
