% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{hess_vec}
\alias{hess_vec}
\title{Compute the Hessian vector product}
\usage{
hess_vec(s_arg1, s_arg2, s_arg3, s_arg4, s_arg5, .copy = FALSE)
}
\arguments{
\item{tag}{integer for tape identification (used in \link{trace_on})}

\item{n}{integer number of independents and m = 1}

\item{x}{independent vector of length n}

\item{v}{vector of length n}

\item{z}{resulting vector of lenth n \nabla^2F(x)}
}
\value{
None. the output is the the parameter z
}
\description{
\code{hess_vec} computes the product of Hessain matrix with a vector of suitable size
}
\examples{
hess_vec(tag,n,x,v,z)

}
\seealso{
{\code{\link{trace_on}}, \code{\link{trace_off}}, 
        \code{\link{badouble_declareIndependent}},
        \code{\link{badouble_declareDependent}}}
}
\keyword{{autodiffadolc}}
