% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{jac_vec}
\alias{jac_vec}
\title{Compute the Jacobian vector product}
\usage{
jac_vec(s_arg1, s_arg2, s_arg3, s_arg4, s_arg5, s_arg6, .copy = FALSE)
}
\arguments{
\item{tag}{integer for tape identification (used in \link{trace_on})}

\item{m}{integer, number of dependents m}

\item{n}{integer, number of independents n}

\item{x}{independent vector x of length n}

\item{v}{vector of length n}

\item{z}{resulting Jacobian vector product of length n}
}
\value{
None. the output is the the parameter z
}
\description{
\code{jac_vec} computes the product of Jacobian matrix with a vector of suitable size
}
\examples{
jac_vec(tag,m,n,x,v,z)

}
\seealso{
{\code{\link{trace_on}}, \code{\link{trace_off}}, 
        \code{\link{badouble_declareIndependent}},
        \code{\link{badouble_declareDependent}}}
}
\keyword{{radolc}}
