% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{jacobian}
\alias{jacobian}
\title{Compute the derivatives of the dependent variables with respect to the independent variables}
\usage{
jacobian(s_arg1, s_arg2, s_arg3, s_arg4, s_arg5, .copy = FALSE)
}
\arguments{
\item{tag}{integer for tape identification (used in \link{trace_on})}

\item{m}{integer, number of dependent variables}

\item{n}{integer number of independents}

\item{x}{independent vector of length n}

\item{J[m][n]}{resulting Jacobian (m*n) F(x)}
}
\value{
None. the output is the the parameter J
}
\description{
\code{jacobian} computes the Jacobian matrix values for m dependents with respect to
n independents for the computation that has been traced. If m>n then the forward mode of automatic differentiation is used. 
Otherwise, the reverse emode of automatic differentiation is used.
}
\examples{
jacobian(tag,m,n,x,J)

}
\seealso{
{\code{\link{trace_on}}, \code{\link{trace_off}}, 
        \code{\link{badouble_declareIndependent}},
        \code{\link{badouble_declareDependent}}}
}
\keyword{{autodiffadolc}}
