% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{trace_off}
\alias{trace_off}
\title{Stop Tracing}
\usage{
trace_off(...)
}
\value{
None
}
\description{
\code{trace_off} ends the \code{radolc} tracing process. Using 
Using \code{radolc} consists of two phases (1) Tracing 
(also called taping), generally performed once and (2) Computation of 
derivatives, which can be performed. \code{trace_off} marks where 
the taping ends.  Cleans up, and turns off trace_flag.
}
\examples{
trace_off()

}
\seealso{
{\code{\link{trace_on}}, \code{\link{badouble_declareIndependent}}, 
        \code{\link{badouble_declareDependent}}}
}
\keyword{@export}
\keyword{{radolc}}
