% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adolc.R
\name{vec_jac}
\alias{vec_jac}
\title{Compute the vector transpose Jacobian vector product}
\usage{
vec_jac(s_arg1, s_arg2, s_arg3, s_arg4, s_arg5, s_arg6, s_arg7,
  .copy = FALSE)
}
\arguments{
\item{tag}{integer for tape identification (used in \link{trace_on})}

\item{m}{integer, number of dependents m}

\item{n}{integer, number of independents n}

\item{x}{independent vector x of length n}

\item{v}{vector of length m}

\item{z}{resulting Jacobian vector product of length n}
}
\value{
None. the output is the the parameter z
}
\description{
\code{vec_jac} computes the product of a transposed vector of suitable size and the Jacobian matrix
}
\examples{
vec_jac(tag,m,n,repeat,x,u,z)

}
\seealso{
{\code{\link{trace_on}}, \code{\link{trace_off}}, 
        \code{\link{badouble_declareIndependent}},
        \code{\link{badouble_declareDependent}}}
}
\keyword{{radolc}}
